Reorganized documents and scripts
