--- ChangeLog
@@ -1,6 +1,0 @@
-See git repository ('git log') for full changelog.
-
-Git repo can be found at:
-https://github.com/kanoi/cgminer
-
-The NEWS file contains most of the changelog
