--- autogen.sh
@@ -1,9 +1,0 @@
-#!/bin/sh
-bs_dir=$(cd "$(dirname "$0")"; pwd)
-
-autoreconf -fi "${bs_dir}"
-
-if test -n "$1" && test -z "$NOCONFIGURE" ; then
-	echo 'Configuring...'
-	"$bs_dir"/configure "$@"
-fi
