+++ docs/AUTHORS
@@ -1,0 +1,9 @@
+Current maintainers and active developers:
+- Raspberry Pi 5: David Means <w1t3h4t+cgminer@gmail.com> bc1qdfnat2vyl9y78rqwu26heua5g9zehjy9kprk0u
+- API+USB+FPGA+ASIC: Andrew Smith <kan0i {at} kano-kun [dot] net> 1Jjk2LmktEQKnv8r2cZ9MvLiZwZ9gxabKm
+- Gekko: https://github.com/vthoang
+
+Legacy:
+- Original CPU mining software: Jeff Garzik <jgarzik@pobox.com>
+- BitFORCE FPGA mining and refactor: Luke Dashjr <luke-jr+cgminer@utopios.org> 1NbRmS6a4dniwHHoSS9v3tEYUpP1Z5VVdL
+- Various code: Con Kolivas <kernel@kolivas.org> 15qSxP1SQcUX3o4nhkfdbgyoWEFMomJ4rZ
