+++ docs/ChangeLog
@@ -1,0 +1,6 @@
+See git repository ('git log') for full changelog.
+
+Git repo can be found at:
+https://github.com/W1T3H4T/cgminer-Raspi
+
+The NEWS file contains most of the changelog
