+++ docs/windows_build/temp/cgminer-gekko-win32/start.bat
@@ -1,0 +1,1 @@
+cgminer.exe -o stratum+tcp://eu.stratum.slushpool.com:3333 -u wareck.gekko -p x --gekko-compac-freq 200 --gekko-2pac-freq 150 --gekko-newpac-freq 150
